refactor: 패키지 명을 cart에서 query로 변경
